adding game controls to docs + clean up
