+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "code-name-dune"
 version = "0.1.0"
-authors = ["Kumar Ankur"]
+authors = ["Kumar Ankur <kmrankur@outlook.com>"]
 edition = "2018"
 
 [dependencies]
