+++ src/systems/mod.rs
@@ -1,7 +1,6 @@
 mod acceleration;
 mod animation;
 mod attack;
-mod camera_motion;
 mod collision;
 mod direction;
 mod motion;
@@ -13,9 +12,9 @@ pub use self::animation::MarineAnimationSystem;
 pub use self::animation::ExplosionAnimationSystem;
 pub use self::animation::PincerAnimationSystem;
 pub use self::attack::AttackSystem;
-pub use self::camera_motion::CameraMotionSystem;
 pub use self::collision::BulletCollisionSystem;
 pub use self::collision::CollisionSystem;
 pub use self::collision::PincerCollisionSystem;
 pub use self::direction::DirectionSystem;
+pub use self::motion::CameraMotionSystem;
 pub use self::motion::MotionSystem;
