+++ src/systems/motion.rs
@@ -1,10 +1,11 @@
 use amethyst::{
     core::Transform,
-    ecs::{Join, ReadStorage, System, WriteStorage},
+    ecs::{Join, ReadExpect, ReadStorage, System, WriteStorage},
 };
 
 use crate::{
-    components::{Collider, Motion, TwoDimObject},
+    components::{Collider, Marine, Motion, Subject, TwoDimObject},
+    resources::Context,
 };
 
 #[derive(Default)]
@@ -35,6 +36,33 @@ impl<'s> System<'s> for MotionSystem {
                 two_dim_obj.set_bottom(collider.next_position.y);
             }
             two_dim_obj.update_transform_position(&mut transform);
+        }
+    }
+}
+
+pub struct CameraMotionSystem;
+
+impl<'s> System<'s> for CameraMotionSystem {
+    type SystemData = (
+        ReadStorage<'s, Marine>,
+        ReadStorage<'s, Subject>,
+        WriteStorage<'s, Transform>,
+        ReadExpect<'s, Context>,
+    );
+
+    fn run(&mut self, (marines, subject_tags, mut transforms, ctx): Self::SystemData) {
+        let mut marine_x = 0.;
+        let map_width = ctx.map_width;
+        let background_width = ctx.bg_width;
+
+        for (_marine, transform) in (&marines, &transforms).join() {
+            marine_x = transform.translation().x.as_f32();
+        }
+
+        for (_subject_tag, transform) in (&subject_tags, &mut transforms).join() {
+            if marine_x >= background_width && marine_x <= map_width - background_width {
+                transform.set_translation_x(marine_x);            
+            }
         }
     }
 }
